+++ storybook/stories/button/Button.template.tsx
@@ -1,5 +1,0 @@
-import { Button, ButtonProps } from '@equinor/fusion-react-button';
-
-export const ButtonTemplate = (props: ButtonProps): JSX.Element => <Button {...props} />;
-
-export default ButtonTemplate;