+++ packages/datepicker/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 0.9.4 (2021-09-16)
+
+**Note:** Version bump only for package @equinor/fusion-react-datepicker
+
+
+
+
+
 ## 0.9.3 (2021-09-08)
 
 **Note:** Version bump only for package @equinor/fusion-react-datepicker